+++ app/src/androidTest/java/org/simple/clinic/drugstockreminders/DrugStockReminderApiIntegrationTest.kt
@@ -2,42 +2,34 @@ package org.simple.clinic.drugstockreminders
 
 import com.google.common.truth.Truth.assertThat
 import org.junit.Before
-import org.junit.Ignore
+import org.junit.Rule
 import org.junit.Test
+import org.junit.rules.RuleChain
 import org.simple.clinic.TestClinicApp
 import org.simple.clinic.drugstockreminders.DrugStockReminder.Result.Found
-import org.simple.sharedTestCode.util.TestUtcClock
-import java.time.LocalDate
-import java.time.LocalDateTime
-import java.time.format.DateTimeFormatter
-import java.util.Locale
+import org.simple.clinic.rules.ServerAuthenticationRule
+import org.simple.sharedTestCode.util.Rules
 import javax.inject.Inject
 
-@Ignore("remove ignore annotation after server implements the drug stock reminder api")
 class DrugStockReminderApiIntegrationTest {
 
   @Inject
   lateinit var drugStockReminder: DrugStockReminder
 
-  @Inject
-  lateinit var clock: TestUtcClock
+  @get:Rule
+  val ruleChain: RuleChain = Rules
+      .global()
+      .around(ServerAuthenticationRule())
 
   @Before
   fun setUp() {
     TestClinicApp.appComponent().inject(this)
-    clock.setDate(LocalDate.parse("2022-03-01"))
   }
 
   @Test
   fun drug_stock_reports_are_filled_or_not_in_the_previous_month_should_be_fetched_to_show_reminders() {
-    // given
-    val dateFormat = DateTimeFormatter.ofPattern("YYYY-MM-DD", Locale.ENGLISH)
-    val previousMonthsDate = LocalDateTime.now(clock).minusMonths(1)
-
-    val formattedDate = dateFormat.format(previousMonthsDate)
-
     // when
-    val result = drugStockReminder.reminderForDrugStock(formattedDate) as Found
+    val result : Found = drugStockReminder.reminderForDrugStock(date = "2022-07-01") as Found
 
     // then
     assertThat(result).isNotNull()